Added diffuse and ambient lighting
Added ambient lighting component. Can be added to any component
and adds light on all sides of meshes. Added diffuse lighting. Only
one allowed in the scene currently and casts directional lighting based
on the position of the mesh and the position of the diffuse component.

Also added debug controls after pressing R